Add files via upload
This is where I scraped additional match specific data to supplement our original dataset. Each row had a url that had to be opened to scrape individual match data. I also formatted the data to be in a more useable structure. The end of this code created the Liga_Arg_Season_2024.csv that is attached.